Extract createTransactionsInExternalVendors logic from index.ts
